[MOV] im_livechat: move ThreadTypingMixin outside of public_livechat.js
Task-2212347
